file operations

+++ PythonCourse/Chap13/types.py
@@ -2,7 +2,7 @@
 # Copyright 2009-2017 BHG http://bw.org/
 
 x = '47'
-y = int(x)
+y = float(x)
 
 print(f'x is {type(x)}')
 print(f'x is {x}')
